Updated Serializer-Deserialiser classes and project
Create a persistent shared (static) variable to hold the serialized XML file classes

Updated the serializer/deserializer classes to make the fields type BindingList.

Now we can use the LoadFromFile method to deserialize and SaveToFile method to serialize.

Set the datasource for the GridControl to the deserialized field class then show  and edit data. When ready SaveToFile to serialize back to XML.